Fix failing test

+++ src/test/kotlin/com/quakbo/euler/Euler8KtTest.kt
@@ -3,6 +3,7 @@ package com.quakbo.euler
 import org.hamcrest.CoreMatchers.`is`
 import org.junit.Assert.assertThat
 import org.junit.Test
+import java.math.BigInteger
 
 class Euler8KtTest {
 
@@ -10,13 +11,13 @@ class Euler8KtTest {
     fun `productOf 12345`() {
         val (product, digits) = productOf("12345")
         assertThat(digits, `is`("12345"))
-        assertThat(product, `is`(120))
+        assertThat(product, `is`(BigInteger.valueOf(120)))
     }
 
     @Test
     fun `productOf 9781797784617`() {
         val (product, digits) = productOf("9781797784617")
         assertThat(digits, `is`("9781797784617"))
-        assertThat(product, `is`(2091059712))
+        assertThat(product, `is`(BigInteger.valueOf(2091059712)))
     }
 }
